+++ src/components/Jobs/AllJobs.js
@@ -107,33 +107,7 @@ const AllJobs = () => {
       });
   };
 
-  const bookMarkPost = (postId) => {
-    axios({
-      method: "post",
-      url: `http://localhost:5000/student/bookmarkJob/${postId}`,
-      headers: {
-        Authorization: "Bearer " + localStorage.getItem("jwt"),
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => {
-        console.log(res);
-        if (res.data.error) {
-          // console.log(res.data.error);
-          const notify = () => toast(res.data.error);
-          notify();
-        } else {
-          // setInternships(res.data.internships);
-          // window.location.reload(false);
-          console.log(res.data.message);
-          const notify = () => toast(res.data.message);
-          notify();
-        }
-      })
-      .catch((err) => {
-        console.log("Error: ", err);
-      });
-  };
+;
 
   return (
     <div className="internshipsOuterContainer">
@@ -173,7 +147,6 @@ const AllJobs = () => {
                 <JobsCard
                   job={job}
                   deletePost={deletePost}
-                  bookMarkPost={bookMarkPost}
                   key={job._id}
                   userId={state.user._id}
                 />
