+++ src/utils/UI/JobsCard/JobsCard.js
@@ -3,9 +3,39 @@ import "../InternshipCard/InternshipCard.css";
 import * as Icon from "react-bootstrap-icons";
 import { Dropdown } from "react-bootstrap";
 import { UserContext } from "../../../App";
+import axios from "axios";
+import toast, { Toaster } from "react-hot-toast";
 
-const JobsCard = ({ job, deletePost, bookMarkPost, userId }) => {
+const JobsCard = ({ job, deletePost, userId }) => {
   const { state, dispatch } = useContext(UserContext);
+
+  const bookMarkPost = (postId) => {
+    axios({
+      method: "post",
+      url: `http://localhost:5000/student/bookmarkJob/${postId}`,
+      headers: {
+        Authorization: "Bearer " + localStorage.getItem("jwt"),
+        "Content-Type": "application/json",
+      },
+    })
+      .then((res) => {
+        console.log(res);
+        if (res.data.error) {
+          // console.log(res.data.error);
+          const notify = () => toast(res.data.error);
+          notify();
+        } else {
+          // setInternships(res.data.internships);
+          // window.location.reload(false);
+          console.log(res.data.message);
+          const notify = () => toast(res.data.message);
+          notify();
+        }
+      })
+      .catch((err) => {
+        console.log("Error: ", err);
+      });
+  };
   const GettingMonth = (date) => {
     const monthNames = [
       "January",
